Use a semaphore to protect dynamodb-local
dynamodb-local is not a high-performance database
and will error (500 response) if too many concurrent
create_table requests are in flight.

- Add a call to DescribeTable in test_transact_write_items

This test seems to be particularly sensitive to the error.
Probably because it is the latest test, not due to anything
specific within the test case.

An early check to test the existance of the table proves this.

- Wrap all calls using the TemporaryDatabase with the semaphore

In particular, acquire all permits when creating the table, and
add a small sleep between requests.
This seems to be sufficient to give dynamodb-local a "rest period".

Other operations seem to not be as expensive, and so we can tolerate
some concurrency in those operations by only acquiring a single permit.